ImagePresentation: extract EntryPoint structs into resource header
